requirements.txt
gTTS==1.2.0
Pillow==3.1.2
xlib==0.21
